pyproject.toml
[tool.poetry]
name = "testing-gcp"
version = "0.1.0"
description = ""
authors = ["azeem <azeem012001@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
google-cloud-translate = "^3.19.0"
summarization = "^0.0.1"
google-cloud = "^0.34.0"
google-cloud-vision = "^3.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
